试卷导入调试成功

+++ do/src/main/java/com/ScoopLink/manageQuestion/analysisQuestion/dto/AnalysisQuestion.java
@@ -38,5 +38,5 @@ public class AnalysisQuestion {
     private Integer sortOrder;
 
 
-    private Integer questionTypeId;
+    private Long questionTypeId;
 }
