+++ do/src/main/java/com/ScoopLink/manageQuestion/essayQuestions/dto/EssayQuestion.java
@@ -40,5 +40,5 @@ public class EssayQuestion {
      /**
      * 题目类型ID
      */
-    private Integer questionTypeId;
+    private Long questionTypeId;
 }
